requirements.txt
Flask==3.1.1
langchain==0.3.25
langchain-community==0.3.25
langchain-openai==0.3.24
openai==1.88.0
faiss-cpu==1.11.0
PyMuPDF==1.26.1
python-dotenv==1.1.0
